+++ lib/src/templates/test/module_templates.dart
@@ -149,6 +149,7 @@ jobs:
       - run: flutter pub get
       - run: dart format --set-exit-if-changed .
       - run: flutter analyze --fatal-infos
+      - run: dart run stackchain doctor --skip-analyze
       - run: flutter test
 ''';
 }